[svc] Fix SvcTest.SecondPassEncode failure
Root cause is the different default register length between x86
and x64 platform. Change spatial_layer_id to long long.

+++ vp9/encoder/vp9_firstpass.h
@@ -35,7 +35,7 @@ typedef struct {
   double new_mv_count;
   double duration;
   double count;
-  int spatial_layer_id;
+  long long spatial_layer_id;
 } FIRSTPASS_STATS;
 
 struct twopass_rc {
